+++ frontend/lib/ui/Form/index.ts
@@ -1,4 +1,3 @@
 export * from "./FormLayout";
 export * from "./Form";
-export * from "./validation";
 export * from "./useFormKeyboardNavigation";
