--- frontend/lib/ui/Form/validation/index.ts
@@ -1,2 +1,0 @@
-export * from "./useForm";
-export * from "./form.util";
