+++ frontend/lib/util/index.ts
@@ -1,12 +1,10 @@
-export * from "./hook/useDebouncedCallback";
-export * from "./hook/useNumericParam";
-export * from "./hook/usePreventFormEnterSubmit";
-export * from "./hook/usePollingStationStatus";
+export * from "./hook";
 export * from "./classnames";
 export * from "./compare";
 export * from "./data-entry";
 export * from "./domtoren";
-export * from "./fields";
-export * from "./strings";
-export * from "./format";
 export * from "./env";
+export * from "./fields";
+export * from "./form";
+export * from "./format";
+export * from "./strings";
